+++ docs/html/search/classes_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['clock_231',['Clock',['../classace__time_1_1clock_1_1Clock.html',1,'ace_time::clock']]]
+  ['clock_247',['Clock',['../classace__time_1_1clock_1_1Clock.html',1,'ace_time::clock']]]
 ];