+++ src/ace_time/ExtendedZone.h
@@ -25,10 +25,6 @@ class ExtendedZone {
     ExtendedZone(const extended::ZoneInfo* zoneInfo):
         mZoneInfoBroker(zoneInfo) {}
 
-    // use default copy constructor and assignment operator
-    ExtendedZone(const ExtendedZone&) = default;
-    ExtendedZone& operator=(const ExtendedZone&) = default;
-
 // TODO: Merge this with BasicZone.h now that they both use the same
 // ACE_TIME_USE_PROGMEM macro.
 #if ACE_TIME_USE_PROGMEM
@@ -59,6 +55,10 @@ class ExtendedZone {
     }
 
   private:
+    // disable copy constructor and assignment operator
+    ExtendedZone(const ExtendedZone&) = delete;
+    ExtendedZone& operator=(const ExtendedZone&) = delete;
+
     const extended::ZoneInfoBroker mZoneInfoBroker;
 };